add shutdown to exercutionmodule
